introduced a new ejb bean for im account management
